+++ Formula/gits.rb
@@ -1,10 +1,10 @@
 class Gits < Formula
   desc "Project-scoped Git submodule workflow with a shared repository cache"
   homepage "https://github.com/leo1394/homebrew-gits"
-  url "https://raw.githubusercontent.com/leo1394/homebrew-gits/v0.1.0/bin/gits", using: :nounzip
-  sha256 "9c99ed0a0c9e86221379a7ef20e7a5e392bddb78210aadfe662f876d7b4384e0"
+  url "https://raw.githubusercontent.com/leo1394/homebrew-gits/v0.2.0/bin/gits", using: :nounzip
+  sha256 "5c517a0d2e26738cf7612d96aa45be654548243d4a5b658819e6134b081830b8"
   license "MIT"
-  head "https://github.com/leo1394/homebrew-gits.git", branch: "main"
+  head "https://github.com/leo1394/homebrew-gits.git", branch: "master"
 
   depends_on "git"
 
@@ -17,8 +17,8 @@ class Gits < Formula
   end
 
   test do
-    assert_match "gits 0.1.0", shell_output("#{bin}/gits --version")
+    assert_match "gits 0.2.0", shell_output("#{bin}/gits --version")
     system "git", "init", "project"
-    assert_match "shared repository: disabled", shell_output("cd project && #{bin}/gits list")
+    assert_match "disabled", shell_output("cd project && #{bin}/gits list")
   end
 end